# swagger.yaml
openapi: 3.0.0

info:
  title: Task API
  version: 1.0.0
  description: Simple CRUD API for managing tasks

servers:
  - url: http://localhost:3000

paths:
  /tasks:
    get:
      summary: Get all tasks
      responses:
        "200":
          description: List of tasks

    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                done:
                  type: boolean
      responses:
        "201":
          description: Task created

  /tasks/search:
    get:
      summary: Search tasks
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Search results

  /tasks/{id}:
    get:
      summary: Get task by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task found
        "404":
          description: Task not found

    put:
      summary: Update a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                done:
                  type: boolean
      responses:
        "200":
          description: Task updated
        "404":
          description: Task not found

    delete:
      summary: Delete a task
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted
        "404":
          description: Task not found